+++ R/get_wms_raster.R
@@ -121,7 +121,7 @@ get_wms_raster <- function(shape,
                               overwrite, version, styles, interactive)
 
    # ensure consistency between the shape's coordinates and those requested
-   shape <- st_make_valid(shape) %>%
+   shape <- st_make_valid(shape) |>
       st_transform(st_crs(crs))
 
    # create needed grid because of 2048 pixel restriction
@@ -186,8 +186,8 @@ grid <- function(shape, resolution, crs) {
 
    nb_pixel_bbox <- nb_pixel_bbox(st_bbox(shape), resolution, crs)
    n_tiles <- as.numeric(ceiling(nb_pixel_bbox/2048))
-   grid <- st_make_grid(shape, n = n_tiles) %>%
-      st_as_sf() %>%
+   grid <- st_make_grid(shape, n = n_tiles) |>
+      st_as_sf() |>
       st_filter(shape, .predicate = st_intersects)
 
    if(st_is_longlat(st_crs(crs))){
@@ -195,7 +195,7 @@ grid <- function(shape, resolution, crs) {
       grid <- st_transform(grid, "CRS:84")
    }
 
-   grid <- grid %>%
+   grid <- grid |>
       st_as_sfc()
 
 }