+++ tests/testthat/get_apicarto_commune_2/apicarto.ign.fr/api/cadastre/commune-5072af.R
@@ -1,9 +1,0 @@
-structure(list(method = "GET", url = "https://apicarto.ign.fr/api/cadastre/commune?geom=%7B%22type%22%3A%22Point%22%2C%22coordinates%22%3A%5B-4.797999999999999%2C47.762%5D%7D&source=PCI", 
-    status_code = 200L, headers = structure(list(`Access-Control-Allow-Origin` = "*", 
-        `Cache-Control` = "private, no-cache, no-store, must-revalidate", 
-        Expires = "-1", Pragma = "no-cache", Vary = "Origin", 
-        `Access-Control-Allow-Credentials` = "true", `Content-Type` = "application/json; charset=utf-8", 
-        `Content-Length` = "147", Date = "Fri, 28 Oct 2022 17:13:31 GMT", 
-        Connection = "keep-alive", `Keep-Alive` = "timeout=5", 
-        `Set-Cookie` = "REDACTED", `Strict-Transport-Security` = "max-age=31536000; includeSubDomains"), class = "httr2_headers"), 
-    body = charToRaw("{\"type\":\"FeatureCollection\",\"features\":[],\"totalFeatures\":0,\"numberMatched\":0,\"numberReturned\":0,\"timeStamp\":\"2022-10-28T17:13:31.024Z\",\"crs\":null}")), class = "httr2_response")