+++ NEWS.md
@@ -5,6 +5,8 @@
 * Apply DRY principle to option("timeout") handling
 * Add "penmarch.shp" as internal data for test and example
 * Better handling of filename saving for `get_wfs()`
+* Adding support for spatial and ecql filter to `get_wfs()`
+* Remove dependecy to magritrr pipe `%>%` by `|>`
 
 # happign 0.1.7
 * Correct bug preventing `get_apicarto_commune` to work with dep and insee code